refactor(web): Move root component under src/core
Splits the generated app/ folder into core/config (bootstrap
providers and routes) and core/root (the root component), and
drops the unused, empty app.scss stylesheet.
