--- apps/web/src/app/app.routes.ts
@@ -1,3 +1,0 @@
-import { Routes } from "@angular/router";
-
-export const routes: Routes = [];
