--- apps/web/src/app/app.ts
@@ -1,12 +1,0 @@
-import { Component, signal } from "@angular/core";
-import { RouterOutlet } from "@angular/router";
-
-@Component({
-    selector: "app-root",
-    imports: [RouterOutlet],
-    templateUrl: "./app.html",
-    styleUrl: "./app.scss",
-})
-export class App {
-    protected readonly title = signal("web");
-}
