+++ apps/web/src/core/config/app.config.ts
@@ -1,10 +1,9 @@
 import {
     ApplicationConfig,
     provideBrowserGlobalErrorListeners,
-} from "@angular/core";
-import { provideRouter } from "@angular/router";
-
-import { routes } from "./app.routes";
+} from '@angular/core';
+import { provideRouter } from '@angular/router';
+import { routes } from './app.routes';
 
 export const appConfig: ApplicationConfig = {
     providers: [provideBrowserGlobalErrorListeners(), provideRouter(routes)],
