+++ apps/web/src/core/config/app.routes.ts
@@ -1,0 +1,3 @@
+import { Routes } from '@angular/router';
+
+export const routes: Routes = [];
