+++ apps/web/src/core/root/root.ts
@@ -1,0 +1,12 @@
+import { ChangeDetectionStrategy, Component, signal } from '@angular/core';
+import { RouterOutlet } from '@angular/router';
+
+@Component({
+    selector: 'app-root',
+    templateUrl: './root.html',
+    changeDetection: ChangeDetectionStrategy.OnPush,
+    imports: [RouterOutlet],
+})
+export class Root {
+    protected readonly title = signal('web');
+}
