+++ apps/web/src/main.ts
@@ -1,5 +1,5 @@
-import { bootstrapApplication } from "@angular/platform-browser";
-import { App } from "./app/app";
-import { appConfig } from "./app/app.config";
+import { appConfig } from '@/core/config/app.config';
+import { Root } from '@/core/root/root';
+import { bootstrapApplication } from '@angular/platform-browser';
 
-bootstrapApplication(App, appConfig).catch((err) => console.error(err));
+bootstrapApplication(Root, appConfig).catch((err) => console.error(err));
